test: product sku service test suite

+++ src/main/java/com/labprog/labprog/services/ProductSkuService.java
@@ -46,9 +46,15 @@ public class ProductSkuService {
         return productSkusRepository.save(productSku);
     }
 
-    public void deleteProductSkuById(UUID id) {
+    public ProductSkus deleteProductSkuById(UUID id) {
         ProductSkus productSku = getProductSkuById(id);
+
+        if (productSku == null) {
+            throw new RuntimeException("Product sku not found with id" + id);
+        }
+
         productSkusRepository.delete(productSku);
+        return productSku;
     }
 
     private void validateProductSku(ProductSkus productSku) {
@@ -59,6 +65,10 @@ public class ProductSkuService {
 
         if (productSku.getQuantity() == null || productSku.getQuantity() <= 0) {
             throw new RuntimeException("Quantity value is not valid");
+        }
+
+        if (productSku.getSku() == null || productSku.getSku().isEmpty()) {
+            throw new RuntimeException("Sku value is not valid");
         }
 
     }
